dhcpcd: point to correct pid file
We moved the pidfile from /var/run to /run.

This should not make a difference, as there should be a
symlink, but in case of buggy setups (such as one of my
computers), this might cause problems due to systemd not
finding the pidfile, timing out, and killing the dhcp daemon.

Signed-off-by: Tom Gundersen <teg@jklm.no>